line_chatbot_api.py
from linebot import (
    LineBotApi, WebhookHandler
)
from linebot.exceptions import (
    InvalidSignatureError
)
from linebot.models import (
    MessageEvent, 
    PostbackEvent,
    TextMessage, 
    TextSendMessage, 
    ImageSendMessage, 
    StickerSendMessage, 
    LocationSendMessage,
    TemplateSendMessage,
    ButtonsTemplate,
    PostbackAction,
    MessageAction,
    URIAction,
    CarouselTemplate,
    CarouselColumn,
    ImageCarouselTemplate,
    ImageCarouselColumn,
    DatetimePickerAction,
    ConfirmTemplate
)

line_bot_api = LineBotApi('5/YbZPl8e6JqCUuKO8P2Spg3NZKmrxQLkx58nris1QfIG6E6fDcfevIkTIZ8Ar3BLA/92xB9vnEKK+wKiSl/kbnjxZIVofSlEjUtwiLPQoZmsY1UJsDQLObauojy6rmGuSDkDCrPXnPXW+r9VeVZHQdB04t89/1O/w1cDnyilFU=')
handler = WebhookHandler('f3b1f35b93482b2df0c111467be05099')